packages/ai-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
openai-whisper==20231117

# Computer Vision
ultralytics>=8.0.200
opencv-python==4.8.1.78
Pillow==10.1.0

# CLIP for scene understanding
open-clip-torch==2.24.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
numpy==1.24.4
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
psutil==5.9.6

# Logging and monitoring
loguru==0.7.2

# Development
pytest==7.4.3
httpx==0.25.2 